test(huggingface): Additional tests and some refactoring.
