+++ tests/conversion/test_configs/gpt2_config_test.yaml
@@ -7,10 +7,10 @@ model:
     sequence_length: 256
     prediction_key: logits
     vocab_size: 50304 # GPT-2 vocab_size of 50257, padded up to nearest multiple of 64 for efficiency
-    n_layer: 6
+    n_layer: 4
     n_head_q: 12
     n_head_kv: 12
-    ffn_hidden: 2048
+    ffn_hidden: 1024
     n_embd: 768
     dropout: 0.0
     bias: false # true # True: bias in Linears, like GPT-2. False: a bit better and faster
